pyproject.toml
[tool.poetry]
name = "ale"
version = "0.1.0"
description = "Auto-Review Tool"
authors = ["Vladyslav Levkovich <levkovich.vlad.2004@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.4.0"
python-dotenv = "^1.0.1"
openai = "^1.52.0"
redis = "^5.1.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

